move and rename files using kebab-case
