+++ src/components/modules/PostCard.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import { Badge } from '@/components/ui/badge';
 import { Image } from 'next-sanity/image';
 import { ArrowRight } from 'lucide-react';
-import ReadTime from '@/components/ReadTime';
+import { ReadTime } from '@/components/ui/read-time';
 import { urlForImage } from '@/lib/sanity/client/utils';
 import { PostCardFragmentType } from '@/lib/sanity/queries/fragments/fragment.types';
 import { getDocumentLink } from '@/lib/links';
