+++ src/components/ui/Date.tsx
@@ -1,6 +1,6 @@
 import { format } from 'date-fns';
 
-export default function DateComponent({ dateString }: { dateString: string | undefined }) {
+export function DateComponent({ dateString }: { dateString: string | undefined }) {
   if (!dateString) {
     return null;
   }
