+++ src/components/ui/read-time.tsx
@@ -1,7 +1,7 @@
 import { Clock } from 'lucide-react';
 import { readTime } from '@/utils/strings';
 
-export default function ReadTime({ wordCount }: { wordCount: number }) {
+export function ReadTime({ wordCount }: { wordCount: number }) {
   return (
     <div className="flex items-center text-sm text-gray-500">
       <Clock className="w-4 h-4 mr-1" />
